リファクタリング(TASK-004: 巨大コンポーネント分割)
